backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.111.0
fastapi-cli==0.0.23
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.2.6
openai==1.30.1
openpyxl==3.1.5
orjson==3.11.7
pandas==2.3.3
pillow==10.3.0
pydantic==2.7.2
pydantic-settings==2.13.1
pydantic_core==2.18.3
Pygments==2.19.2
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
rich==14.3.3
rich-toolkit==0.19.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.4.0
tqdm==4.67.3
typer==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
faiss-cpu

python-multipart
sse-starlette

rank_bm25==0.2.2
